+++ src/NumericHelper.php
@@ -16,12 +16,45 @@ use function is_numeric;
 use function is_scalar;
 use function preg_replace;
 use function str_replace;
+use function substr;
 
 /**
  * Provides static methods to work with numeric strings.
  */
 final class NumericHelper
 {
+    /**
+     * @psalm-var array<int, array<string, int>>
+     */
+    private const FILESYSTEM_SIZE_POSTFIXES = [
+        3 => [
+            'KiB' => 1024,
+            'MiB' => 1048576,
+            'GiB' => 1073741824,
+            'TiB' => 1099511627776,
+            'PiB' => 1125899906842624,
+        ],
+        2 => [
+            'kB' => 1000,
+            'MB' => 1000000,
+            'GB' => 1000000000,
+            'TB' => 1000000000000,
+            'PB' => 1000000000000000,
+        ],
+        1 => [
+            'k' => 1024,
+            'K' => 1024,
+            'm' => 1048576,
+            'M' => 1048576,
+            'g' => 1073741824,
+            'G' => 1073741824,
+            't' => 1099511627776,
+            'T' => 1099511627776,
+            'p' => 1125899906842624,
+            'P' => 1125899906842624,
+        ],
+    ];
+
     /**
      * Converts number to its ordinal English form. For example, converts 13 to 13th, 2 to 2nd etc.
      *
@@ -85,5 +118,47 @@ final class NumericHelper
     public static function isInteger(mixed $value): bool
     {
         return filter_var($value, FILTER_VALIDATE_INT) !== false;
+    }
+
+    /**
+     * Converts human readable size to bytes.
+     *
+     * @param string $string Human readable size. Examples: `1024`, `1kB`, `1.5M`, `1GiB`. Full
+     * list of supported postfixes in {@see FILESYSTEM_SIZE_POSTFIXES}.
+
+     * Note: This parameter must be less than `8192P` on 64-bit systems and `2G` on 32-bit systems.
+     *
+     * @throws InvalidArgumentException when the string is invalid.
+     *
+     * @return int The number of bytes equivalent to the specified string.
+     *
+     * @see https://www.gnu.org/software/coreutils/manual/html_node/Block-size.html
+     */
+    public static function convertHumanReadableSizeToBytes(string $string): int
+    {
+        if (is_numeric($string)) {
+            return (int) $string;
+        }
+
+        foreach (self::FILESYSTEM_SIZE_POSTFIXES as $postfixLength => $postfixes) {
+            $postfix = substr($string, -$postfixLength);
+            if ($postfix === '' || preg_match('/\\d/', $postfix) === 1) {
+                continue;
+            }
+
+            $numericPart = substr($string, 0, -$postfixLength);
+            if (!is_numeric($numericPart)) {
+                throw new InvalidArgumentException("Incorrect input string: $string");
+            }
+
+            $postfixMultiplier = $postfixes[$postfix] ?? null;
+            if ($postfixMultiplier === null) {
+                throw new InvalidArgumentException("Not supported postfix '$postfix' in input string: $string");
+            }
+
+            return (int) ((float) $numericPart * $postfixMultiplier);
+        }
+
+        throw new InvalidArgumentException("Incorrect input string: $string");
     }
 }